Fix a bunch of errors, delete firebase
