--- client/.firebaserc
@@ -1,5 +1,0 @@
-{
-  "projects": {
-    "default": "math-tts-464ea"
-  }
-}
