+++ client/package.json
@@ -13,11 +13,11 @@
     "axios": "^1.2.2",
     "react": "^18.2.0",
     "react-audio-player": "^0.17.0",
-    "react-awesome-spinners": "^1.3.1",
     "react-bootstrap": "^2.7.0",
     "react-dom": "^18.2.0",
     "react-dotenv": "^0.1.3",
     "react-scripts": "5.0.1",
+    "react-spinners": "^0.13.7",
     "typescript": "^4.9.4",
     "web-vitals": "^2.1.4",
     "xstate": "^4.35.2"
